handsOn_codes/01- FastAPI and a simple deep model/requirement.txt
ultralytics==8.0.42
uvicorn[standard]==0.20.0
gunicorn==20.1.0
fastapi
loguru~=0.7.0
pytest~=7.4.0
pytest-cov
pandas~=2.0.3
pillow~=10.0.0
requests~=2.31.0
fastapi~=0.101.1
numpy~=1.25.2
torch~=2.0.1
torchvision~=0.15.2